little bugfix

+++ mainwindow.cpp
@@ -206,8 +206,8 @@ void MainWindow::on_pushButton_3_clicked()
                     size_t usedMemory = (scti(recived[4]) << 24) | (scti(recived[5]) << 16) | (scti(recived[6]) << 8) | scti(recived[7]);
                     ui->memory->setRange(0, totalMemory);
                     ui->memory->setValue(totalMemory - usedMemory);
-                    ui->totalMemory->setText(QString::number((int32_t)totalMemory));
-                    ui->availableMemory->setText(QString::number((int32_t)(totalMemory - usedMemory)));
+                    ui->totalMemory->setText(QString::number(totalMemory));
+                    ui->availableMemory->setText(QString::number((totalMemory - usedMemory)));
                     recived.clear();
                     ui->listWidget->clear();
                     mode = "LIST";
